better component collection
findAllOfType template function added to the edge collector:
an issue is fixed where only the first layer of components is checked, which is now recursive for any aactor holding any other aactors. Used here to collect all UStaticMeshes. The class / Type which may be searched for must be derived of UActorComponent.